requirements.txt
streamlit>=1.30.0
faster-whisper>=1.0.0
pyannote.audio>=3.1.0
python-docx>=1.1.0
weasyprint>=60.0
torch>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
